+++ notification/src/epilot/models/operations/createnotification.py
@@ -6,7 +6,6 @@ import requests as requests_http
 from typing import Optional
 
 
-
 @dataclasses.dataclass
 class CreateNotificationResponse:
     content_type: str = dataclasses.field()