+++ notification/src/epilot/models/operations/getnotification.py
@@ -3,8 +3,8 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from typing import Any, Optional
-
+from ..shared import notificationitem as shared_notificationitem
+from typing import Optional
 
 
 @dataclasses.dataclass
@@ -15,14 +15,13 @@ class GetNotificationRequest:
 
 
 
-
 @dataclasses.dataclass
 class GetNotificationResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    notification_item: Optional[dict[str, Any]] = dataclasses.field(default=None)
+    notification_item: Optional[shared_notificationitem.NotificationItem] = dataclasses.field(default=None)
     r"""Success"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""